finish bank side

+++ bank-management/bank-core/src/main/java/com/example/bankcore/events/AccountOpenedEvent.java
@@ -23,6 +23,6 @@ public class AccountOpenedEvent {
 
     private Date createDate;
 
-    private BigDecimal openingBalance;
+    private double openingBalance;
 
 }
