+++ bank-management/bank-core/src/main/java/com/example/bankcore/events/FundsDepositedEvent.java
@@ -15,9 +15,9 @@ public class FundsDepositedEvent {
 
     private String id;
 
-    private BigDecimal amount;
+    private double amount;
 
-    private BigDecimal balance;
+    private double balance;
 
 
 }
