+++ bank-management/bank-core/src/main/java/com/example/bankcore/events/FundsWithdrawEvent.java
@@ -15,8 +15,8 @@ public class FundsWithdrawEvent {
 
     private String id;
 
-    private BigDecimal amount;
+    private double amount;
 
-    private BigDecimal balance;
+    private double balance;
 
 }
